Rearrange python2js_buffer body into a js file
